Create AbstractAuthHandler to fix final circular dep
