+++ packages/imperative/index.ts
@@ -22,5 +22,6 @@ export * from "./src/DefinitionTreeResolver";
 export * from "./src/Imperative";
 export * from "./src/LoggingConfigurer";
 export * from "./src/plugins/PluginManagementFacility";
+export * from "./src/auth/handlers/AbstractAuthHandler";
 export * from "./src/auth/handlers/BaseAuthHandler";
 export * from "./src/config/cmd/auto-init/handlers/BaseAutoInitHandler";
